refactor(recipe-import): update RecipeImportCompleted event to use Recipe model
Modified the RecipeImportCompleted event to accept a Recipe model instead of a recipe ID. Updated the ImportRecipeJob to dispatch the event with the Recipe url, enhancing the data passed to the frontend. Adjusted the frontend composable to utilize the new recipe URL and structure for improved navigation and user experience.